move API load testing with k6
