+++ .gitignore
@@ -1,6 +1,7 @@
 node_modules
 .DS_Store
 .idea
+.ignore
 !.vscode/settings.json
 .env
 db/*
