+++ api/providers/geo/src/providers/EtalabGeoAdressProvider.ts
@@ -14,7 +14,7 @@ export class EtalabGeoAdressProvider implements GeoCoderInterface, InseeCoderInt
     axiosRetry(axios, {
       retries: 3,
       retryDelay: (retryCount) => retryCount * 2000,
-      retryCondition: (error) => error.response.status >= 400,
+      retryCondition: (error) => (error.response?.status || 500) >= 400,
     });
   }
